+++ tests/run.php
@@ -1,5 +1,6 @@
 <?php
 
+require __DIR__ . '/../vendor/autoload.php';
 require __DIR__ . '/../src/autoload.php';
 require __DIR__ . '/TestCase.php';